take main method out of public API
